WIP: Improve handling of secret permissions updates
Add a mutex for policy updates; updates to resources can happen
concurrently, and we have issues if we try to do policy updates on the
same branch at the same time.

Whilst we're here, also check if the permissions have changed and only
send a policy update if they have - hopefully reduce the number of locks
we need to have.